requirements/base.txt
mistune
pygments
django-reversion
django-compressor
django-pagedown
